+++ CHANGELOG.rst
@@ -10,6 +10,9 @@ Unreleased
 
 Added
 -----
+* A submodule `results` that implements functionality to visualise or format
+  the results of one or more experiments into neat little plots and tables.
+  Accompanying this new submodule is a new example script.
 * The Experiment class how is an abstract class with abstract methods that
   function as events. This allows users to more easily create Experiment
   classes that extend the functionality of the Experiment class.